DO NOT MERGE: Initial work on decoupling prompting.
This commit is not complete. The work on the service package is, and the
work on the cmd/mutagen/daemon package is, but commands have not yet
been transitioned to use decoupled prompting. This is turning out to be
a little more tedious than anticipated, and there are certain aspects of
it (e.g. monitoring for prompting failure, which would require the
actual method invocation to happen in a background Goroutine as well)
that make it sort of ugly.

Anyway, the aforementioned completed packages should be good to go, it's
just the commands that need to be transitioned.

I'm thinking that something like a global daemon connection and/or
global prompter might help, but managing their lifecycles and errors is
really difficult to do elegantly.

Signed-off-by: Jacob Howard <jacob@havoc.io>